requirements.txt
opencv-python
keyboard
vgamepad
pillow
numpy
black
mss